requirements.txt
requests
networkx
pandas
matplotlib
python-louvain
tqdm
python-dotenv
scipy
